Extract AbstractGenericService to de-duplicate service base classes
GenericService and GenericKtorService were ~95% identical: the same constructor
parameters, fields, Prometheus/Zipkin/health-check wiring, ServiceManager setup,
start/stop lifecycle, and shutDownHookAction. Only the admin servlet-hosting
backend differed (Jetty ServletService+ServletGroup vs Ktor KtorServletService+
HttpServletGroup, plus the extra initKtor block).

Move the shared machinery into a new AbstractGenericService<T> base class. Each
subclass now holds only its servlet-hosting specifics and exposes its servlet
service to the shared lifecycle via the servletServiceOrNull hook. Public API is
unchanged: configVals, isTestMode, the isXEnabled flags, jmxReporter/
metricsService/zipkinReporterService, upTime, the typed servletService, the
init* methods, and GenericService.shutDownHookAction (kept as a thin delegate)
all remain accessible exactly as before.

The only observable change is that the shared log lines now log under the
AbstractGenericService logger category instead of the per-subclass category.

Adds GenericServiceTests, which subclasses both variants and pins the shared
health-check registration, enablement flags, and lifecycle wiring.
